# docs/swagger.json
{
    "consumes": [
        "application/json"
    ],
    "produces": [
        "application/json"
    ],
    "swagger": "2.0",
    "info": {
        "description": "Данный сервис предоставляент возможность поиска информации об адресе, по его фактическом адресу или геолокации.",
        "title": "Поиск данных адреса",
        "contact": {
            "name": "Вячеслав",
            "email": "slavalomov.4@gmail.com"
        },
        "version": "1.0"
    },
    "host": "localhost:8080",
    "basePath": "/",
    "paths": {
        "/api/address/geocode": {
            "post": {
                "description": "Поиск информации об адресе по его геолокационным данным.",
                "summary": "Получение данных адреса.",
                "parameters": [
                    {
                        "description": "координаты",
                        "name": "input",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/models.GeocodeRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "информация об адресе",
                        "schema": {
                            "$ref": "#/definitions/models.GeocodeResponse"
                        }
                    },
                    "400": {
                        "description": "400 :Неверный формат запроса",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "500": {
                        "description": "500: Сервис dadata.ru не доступен",
                        "schema": {
                            "type": "string"
                        }
                    }
                }
            }
        },
        "/api/address/search": {
            "post": {
                "description": "Поиск информации об адресе по его фактическому адресу",
                "summary": "Получение данных адреса",
                "parameters": [
                    {
                        "description": "фактический адрес",
                        "name": "input",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/models.SearchRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "информация об адресе",
                        "schema": {
                            "$ref": "#/definitions/models.SearchResponse"
                        }
                    },
                    "400": {
                        "description": "400 :Неверный формат запроса",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "500": {
                        "description": "500: Сервис dadata.ru не доступен",
                        "schema": {
                            "type": "string"
                        }
                    }
                }
            }
        }
    },
    "definitions": {
        "models.Address": {
            "type": "object",
            "properties": {
                "city": {
                    "type": "string"
                },
                "geo_lat": {
                    "type": "string"
                },
                "geo_lon": {
                    "type": "string"
                },
                "house": {
                    "type": "string"
                },
                "street": {
                    "type": "string"
                }
            }
        },
        "models.GeocodeRequest": {
            "type": "object",
            "properties": {
                "lat": {
                    "type": "number"
                },
                "lon": {
                    "type": "number"
                }
            }
        },
        "models.GeocodeResponse": {
            "type": "object",
            "properties": {
                "addresses": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/models.Address"
                    }
                }
            }
        },
        "models.SearchRequest": {
            "type": "object",
            "properties": {
                "query": {
                    "type": "string"
                }
            }
        },
        "models.SearchResponse": {
            "type": "object",
            "properties": {
                "addresses": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/models.Address"
                    }
                }
            }
        }
    }
}